Refactoring: moving PageMethods methods to DefaultSteps
